+++ tests/Unit/Requests/Payments/SBP/SBPQrLinkRequestBuilderTest.php
@@ -6,12 +6,11 @@ namespace AvtoDev\Tests\Unit\Requests\Payments\SBP;
 
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
+use PHPUnit\Framework\Attributes\CoversClass;
 use AvtoDev\Tests\Unit\Requests\AbstractRequestBuilderTestCase;
 use AvtoDev\CloudPayments\Requests\Payments\SBP\SBPQrLinkRequestBuilder;
 
-/**
- * @covers \AvtoDev\CloudPayments\Requests\Payments\SBP\SBPQrLinkRequestBuilder
- */
+#[CoversClass(SBPQrLinkRequestBuilder::class)]
 class SBPQrLinkRequestBuilderTest extends AbstractRequestBuilderTestCase
 {
     /**