+++ tests/Unit/Requests/Payments/Tokens/TokensAuthRequestBuilderTest.php
@@ -6,12 +6,11 @@ namespace AvtoDev\Tests\Unit\Requests\Payments\Tokens;
 
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
+use PHPUnit\Framework\Attributes\CoversClass;
 use AvtoDev\Tests\Unit\Requests\AbstractRequestBuilderTestCase;
 use AvtoDev\CloudPayments\Requests\Payments\Tokens\TokensAuthRequestBuilder;
 
-/**
- * @covers \AvtoDev\CloudPayments\Requests\Payments\Tokens\TokensAuthRequestBuilder
- */
+#[CoversClass(TokensAuthRequestBuilder::class)]
 class TokensAuthRequestBuilderTest extends AbstractRequestBuilderTestCase
 {
     /**