+++ tests/Unit/Requests/Subscriptions/SubscriptionsCreateRequestBuilderTest.php
@@ -9,12 +9,11 @@ use GuzzleHttp\Psr7\Uri;
 use Illuminate\Support\Str;
 use Psr\Http\Message\UriInterface;
 use AvtoDev\CloudPayments\Receipts\Receipt;
+use PHPUnit\Framework\Attributes\CoversClass;
 use AvtoDev\Tests\Unit\Requests\AbstractRequestBuilderTestCase;
 use AvtoDev\CloudPayments\Requests\Subscriptions\SubscriptionsCreateRequestBuilder;
 
-/**
- * @covers \AvtoDev\CloudPayments\Requests\Subscriptions\SubscriptionsCreateRequestBuilder
- */
+#[CoversClass(SubscriptionsCreateRequestBuilder::class)]
 class SubscriptionsCreateRequestBuilderTest extends AbstractRequestBuilderTestCase
 {
     /**